Renamed all js files to jsx and resolved related jsx issues
