+++ .eslintrc.json
@@ -1,3 +1,25 @@
 {
-  "extends": ["next/core-web-vitals", "airbnb", "airbnb/hooks"]
+  "extends": ["next/core-web-vitals", "airbnb", "airbnb/hooks"],
+  "rules": {
+    "jsx-a11y/no-static-element-interactions": [
+      "error",
+      {
+        "handlers": ["onClick"],
+        "allowExpressionValues": true,
+        "allowStaticElements": ["div", "span"] // Add any other static elements as needed
+      }
+    ],
+    "import/extensions": [
+      "error",
+      "never"
+    ]
+  },
+  "overrides": [
+    {
+      "files": ["*.js", "*.jsx", "*.ts", "*.tsx"],
+      "rules": {
+        "import/no-unresolved": ["error", { "ignore": ["^@"] }]
+      }
+    }
+  ]
 }
