+++ src/components/NewsArticle/RiskRating.jsx
@@ -70,7 +70,8 @@ function RiskRating({ riskRating }) {
         <p className="text-center text-gray-900 text-xs">
           <span className="text-sm font-semibold riskRating">
             {riskRating.toUpperCase()}
-          </span>{' '}
+          </span>
+          {' '}
           <span className="block text-xs max-w-full overflow-hidden overflow-ellipsis whitespace-nowrap">
             RISK
           </span>
