+++ src/components/Search/SmallSearchBar.jsx
